Implementation (Week 11)/driverless_system.py
class DriverlessSystem:
    def __init__(self):
        self.camera = Camera()
        self.obstacle = Obstacle()
        self.decision = Decision()
        self.data = Data()

    def make_decision(self):
        # Logic for making decisions based on captured data
        if self.obstacle.detected:
            self.decision.result = 1  # Set decision to proceed
        else:
            self.decision.result = 0  # Set decision to halt

    def detect_obstacle(self):
        # Logic for detecting obstacles using camera
        self.camera.capture_image()
        # Assume obstacle detection logic based on captured image
        # Randomly generate detection result, it will have to be replaced to be implemented in real system
        import random
        self.obstacle.detected = random.choice([0, 1])  # Randomly set to 0 or 1

    def store_data(self):
        # Store captured image data in the Data class
        image_data = self.camera.capture_image()
        self.data.store_data(image_data)

class Camera:
    def __init__(self):
        self.capture = None

    def capture_image(self):
        # Logic for capturing image
        # It returns placeholder image, which needs to be replaced with image API if required
        return "Placeholder image"


class Data:
    def __init__(self):
        self.raw_data = None

    def store_data(self, image_data):
        # Store the captured image data
        self.raw_data = image_data


class Obstacle:
    def __init__(self):
        self.detected = 0

class Decision:
    def __init__(self):
        self.result = 0